+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2018-08-22
+
+Bump v2.0.0 :tada:
+
 ### Changed
 - Code cleanup, minor enhancements.